expire session

+++ src/DistribuceTisku/Bundle/Controller/FormController.php
@@ -9,19 +9,13 @@ use DistribuceTisku\Bundle\Entity\Supplier;
 use DistribuceTisku\Bundle\Form\BookType;
 use DistribuceTisku\Bundle\Form\SupplierType;
 use DistribuceTisku\Bundle\Controller\SecurityController;
+use DistribuceTisku\Bundle\Controller\FormController;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 
-class FormController extends Controller {
-    
-    function securityExpire(){
-        
-        $pom = new SecurityController;
-        return $pom->timeCheck();
-        
-    }
+class FormController extends UpperController {
 
-    function getBookByISSN($id) {
-        
+    public function getBookByISSN($id) {
+        if(($pom = $this->timeCheck()) != "") {echo $pom;return $this->render($pom);}
         $b = new Book();
         $conn = $this->get('database_connection');
         $sql = "SELECT * FROM tiskovina WHERE ISSN='" . $id . "';";
@@ -38,7 +32,7 @@ class FormController extends Controller {
     }
 
     public function bookAddAction() {
-        
+        if(($pom = $this->timeCheck()) != "") {echo $pom;return $this->render($pom);}
         $book = new Book();
         $form = $this->createForm(new BookType(), $book);
 
@@ -64,6 +58,7 @@ class FormController extends Controller {
     }
 
     public function bookEditAction($id) {
+        if(($pom = $this->timeCheck()) != "") {echo $pom;return $this->render($pom);}
         $book = $this->getBookByISSN($id);
         $form = $this->createForm(new BookType(), $book);
         $request = $this->getRequest();
@@ -84,6 +79,7 @@ class FormController extends Controller {
     }
 
     public function bookListAction() {
+        if(($pom = $this->timeCheck()) != "") {echo $pom;return $this->render($pom);}
         $request = $this->getRequest();
         if ($request->getMethod() == 'POST') {
             
@@ -95,6 +91,7 @@ class FormController extends Controller {
     }
 
     public function supplierAddAction() {
+        if(($pom = $this->timeCheck()) != "") {echo $pom;return $this->render($pom);}
         $supplier = new Supplier();
         $form = $this->createForm(new SupplierType(), $supplier);
 
@@ -133,8 +130,8 @@ class FormController extends Controller {
         ));
     }
     
-    public function supplierEditAction($id)
-    {        
+    public function supplierEditAction($id)      
+    {        if(($pom = $this->timeCheck()) != "") {echo $pom;return $this->render($pom);}
         $suplier = new Supplier();
         $conn = $this->get('database_connection');
         $sql = "SELECT * FROM `dodavatel` WHERE id_dodavatele = '" . $id . "'";
@@ -167,6 +164,7 @@ class FormController extends Controller {
     
     public function supplierListAction()
     {
+        if(($pom = $this->timeCheck()) != "") {echo $pom;return $this->render($pom);}
         $conn = $this->get('database_connection');
         $sql = "SELECT * FROM `dodavatel`";
         $dodavatel = $conn->prepare($sql);
@@ -176,6 +174,7 @@ class FormController extends Controller {
     
     public function supplierDeleteAction($id)
     {
+        if(($pom = $this->timeCheck()) != "") {echo $pom;return $this->render($pom);}
         $conn = $this->get('database_connection');
         $conn->delete('dodavatel', array('id_dodavatele' => $id));
         return $this->redirect($this->generateUrl('_supplierList'));
@@ -222,6 +221,7 @@ class FormController extends Controller {
     }
 
     public function subscriptionAddAction() {
+        if(($pom = $this->timeCheck()) != "") {echo $pom;return $this->render($pom);}
         $subscription = new Subscription();
         $subscription->setOdberDo(2013);
         $form = $this->makeSubscriptionForm($subscription);
@@ -247,6 +247,7 @@ class FormController extends Controller {
     }
 
     public function subscriptionEditAction($id) {
+        if(($pom = $this->timeCheck()) != "") {echo $pom;return $this->render($pom);}
         $subscription = new Subscription();
         $conn = $this->get('database_connection');
         $sql = "SELECT * FROM `odber` WHERE id_odberu = '" . $id . "'";
@@ -280,7 +281,8 @@ class FormController extends Controller {
         ));
     }
 
-    private function getSubscriptionListAction($id = "") {
+    private function getSubscriptionList($id = "") {
+        if(($pom = $this->timeCheck()) != "") {echo $pom;return $this->render($pom);}
         $conn = $this->get('database_connection');
         $sql = "SELECT `platby`.`obdobi`,`platby`.`zpusob_platby`,`odber`.`id_odberu`, `odber`.`den_odberu`,`odber`.`odber_od`,`odber`.`odber_do`,`odber`.`id_platby`,`zakaznik`.`jmeno`,`zakaznik`.`prijmeni`,`tiskovina`.`titul`  FROM `odber`";
         $sql = $sql . "JOIN `zakaznik` ON `odber`.`id_zakaznika` = `zakaznik`.`id_zakaznika`";
@@ -296,11 +298,13 @@ class FormController extends Controller {
     }
 
     public function subscriptionListAction() {
+        if(($pom = $this->timeCheck()) != "") {echo $pom;return $this->render($pom);}
         $odbery = $this->getSubscriptionList();
         return $this->render('DistribuceTiskuBundle:Form:subscriptionList.html.twig', array('odbery' => $odbery));
     }
 
     public function subscriptionInterruptionAction($id) {
+        if(($pom = $this->timeCheck()) != "") {echo $pom;return $this->render($pom);}
         $interuption = new SubscriptionInterruption();
         $interuption->setId($id);
         $form = $this->createFormBuilder($interuption)
@@ -339,6 +343,7 @@ class FormController extends Controller {
     }
     
     public function subscriptionRemoveAction($id){
+        if(($pom = $this->timeCheck()) != "") {echo $pom;return $this->render($pom);}
         $conn = $this->get('database_connection');
         $conn->delete('odber', array('id_odberu' => $id));
         $conn->delete('preruseni_odberu', array('id_odberu' => $id));
@@ -346,6 +351,7 @@ class FormController extends Controller {
     }
 
     public function subscriptionInterruptionListAction( $id) {
+        if(($pom = $this->timeCheck()) != "") {echo $pom;return $this->render($pom);}
         $conn = $this->get('database_connection');
         $sql = "SELECT `id_preruseni`, `preruseni_od`, `preruseni_do` FROM `preruseni_odberu` WHERE `id_odberu` = '" . $id . "'";
         $preruseni = $conn->prepare($sql);
@@ -354,6 +360,7 @@ class FormController extends Controller {
     }
 
     public function subscriptionInteruptionDeleteAction($id) {
+        if(($pom = $this->timeCheck()) != "") {echo $pom;return $this->render($pom);}
         $conn = $this->get('database_connection');
 
         $conn->delete('preruseni_odberu', array('id_preruseni' => $id));
@@ -369,18 +376,21 @@ class FormController extends Controller {
     //******************USER********************************************************
 
     public function ubookListAction() {
+        if(($pom = $this->timeCheck()) != "") {echo $pom;return $this->render($pom);}
         $conn = $this->get('database_connection');
         $books = $conn->fetchAll('SELECT * FROM tiskovina ORDER BY titul ASC');
         return $this->render('DistribuceTiskuBundle:Form:ubooklist.html.twig', array('books' => $books));
     }
 
     public function usubscriptionListAction() {
+        if(($pom = $this->timeCheck()) != "") {echo $pom;return $this->render($pom);}
         $session = $this->getRequest()->getSession();
         $odbery = $this->getSubscriptionList($session->get('id'));
         return $this->render('DistribuceTiskuBundle:Form:usubscriptionList.html.twig', array('odbery' => $odbery));
     }
 
     public function usubscriptionAddAction() {
+        if(($pom = $this->timeCheck()) != "") {echo $pom;return $this->render($pom);}
         $subscription = new Subscription();
         $subscription->setOdberDo(2013);
         $form = $this->makeSubscriptionForm($subscription);
@@ -406,6 +416,7 @@ class FormController extends Controller {
     }
     
     public function usubscriptionEditAction($id) {
+        if(($pom = $this->timeCheck()) != "") {echo $pom;return $this->render($pom);}
         $subscription = new Subscription();
         $conn = $this->get('database_connection');
         $sql = "SELECT * FROM `odber` WHERE id_odberu = '" . $id . "'";
